+++ src/main/java/com/guildedrose/shop/BalanceRepository.java
@@ -1,6 +1,6 @@
 package com.guildedrose.shop;
 
 public interface BalanceRepository {
-    int GetBalance();
-    void SaveBalance(int balance);
+    double GetBalance();
+    void SaveBalance(double balance);
 }
